Added myexp function and tested to make sure it works

+++ LectureAssignments/Lecture08/lab8.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <stdlib.h>
 int factorial(int x);
-int myexp(int x);
+double myexp(double x);
